+++ tests/test_psycopg2.py
@@ -7,22 +7,31 @@ conn.autocommit = True
 
 cur = conn.cursor()
 cur.execute('CREATE EXTENSION IF NOT EXISTS vector')
-cur.execute('DROP TABLE IF EXISTS items')
-cur.execute('CREATE TABLE items (id bigserial PRIMARY KEY, embedding vector(3))')
+cur.execute('DROP TABLE IF EXISTS psycopg2_items')
+cur.execute('CREATE TABLE psycopg2_items (id bigserial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3))')
 
 register_vector(cur)
 
 
 class TestPsycopg2:
     def setup_method(self, test_method):
-        cur.execute('DELETE FROM items')
+        cur.execute('DELETE FROM psycopg2_items')
 
-    def test_works(self):
+    def test_vector(self):
         embedding = np.array([1.5, 2, 3])
-        cur.execute('INSERT INTO items (embedding) VALUES (%s), (NULL)', (embedding,))
+        cur.execute('INSERT INTO psycopg2_items (embedding) VALUES (%s), (NULL)', (embedding,))
 
-        cur.execute('SELECT * FROM items ORDER BY id')
+        cur.execute('SELECT * FROM psycopg2_items ORDER BY id')
         res = cur.fetchall()
         assert np.array_equal(res[0][1], embedding)
         assert res[0][1].dtype == np.float32
+        assert res[1][1] is None
+
+    def test_halfvec(self):
+        embedding = [1.5, 2, 3]
+        cur.execute('INSERT INTO psycopg2_items (half_embedding) VALUES (%s), (NULL)', (embedding,))
+
+        cur.execute('SELECT id, half_embedding FROM psycopg2_items ORDER BY id')
+        res = cur.fetchall()
+        assert res[0][1].to_list() == [1.5, 2, 3]
         assert res[1][1] is None