+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.6 (unreleased)
 
 - Added support for `halfvec` and `sparsevec` types to Psycopg 3
+- Added support for `halfvec` type to Psycopg 2
 - Added support for `halfvec` and `sparsevec` types to asyncpg
 - Added support for `halfvec` and `sparsevec` types to Peewee
 - Added `L1Distance` for Django